requirements.txt
# Framework Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3

# Serialização e Validação
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Documentação API
apispec==6.3.0
apispec-webframeworks==0.5.2
Flasgger==0.9.7.1

# Banco de Dados
psycopg2-binary==2.9.7

# Utilitários
python-dotenv==1.0.0
reportlab
selenium
